Put separate functions in dom updats file and scripts file
