+++ src/recipe-functions.js
@@ -15,11 +15,11 @@ const findRecipe = (type, recipeList, tag) => {
 
 // Return a specific recipe based on the id number
 
-const specificRecipe = () => {
-  return recipeData.find(recipe => recipe.id === clickedId)
-}
+// const specificRecipe = () => {
+//   return recipeData.find(recipe => recipe.id === clickedId)
+// }
 
 export {
   findRecipe,
-  specificRecipe
+ // specificRecipe
 }
